backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
django==5.2.6
django-cors-headers==4.9.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework-stubs==3.16.4
humanfriendly==10.0
idna==3.11
kombu==5.5.4
nodeenv==1.9.1
numpy==2.3.3
packaging==25.0
pandas==2.3.2
prompt-toolkit==3.0.52
pyright==1.1.406
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
types-pyyaml==6.0.12.20250915
types-requests==2.32.4.20250913
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
